+++ js/main.js
@@ -4,6 +4,7 @@ import './map.js'
 import { renderPins } from './map.js'
 import { getData } from './api.js'
 import { showAlert } from './util.js'
+import './message.js'
 
 const ERROR_GETTING_DATA = 'Не удалось получить данные с сервера. Попробуйте позже';
