+++ js/util.js
@@ -53,5 +53,8 @@ const showAlert = (message) => {
   }, ALERT_SHOW_TIME);
 }
 
+const isEscEvent = (evt) => {
+  return evt.key === ('Escape' || 'Esc');
+};
 
-export { getRandomIntInclusive, getRandomDecimalNumber, getRandomArrayElement, getRandomElementsFromArray, showAlert };
+export { getRandomIntInclusive, getRandomDecimalNumber, getRandomArrayElement, getRandomElementsFromArray, showAlert, isEscEvent };
